add chroma db and refactor helpers into their own directory
